requirements.txt
celery-with-redis==3.0
Django==2.1.9
django-celery==3.3.1
django-cors-headers==3.0.2
django-crontab==0.7.1
django-easy-pdf==0.1.1
django-filter==2.1.0
django-haystack==2.8.1
django-mssql==1.8
django-object-tool==0.0.8
django-pdfkit==0.3.1
django-pytds==1.5
django-sqlserver==1.11
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-xml==1.4.0
drf-extensions==0.5.0
drf-writable-nested==0.5.1
Markdown==3.1.1
django-rq
opencv-python
requests
python-dateutil
django-celery
celery==4.4.7
redis==2.10.6
